requirements.txt
boto3
pandas
alive-progress
python-dotenv